Created an apibase class for accessing the api.
